backend/requirements.txt
fastapi
uvicorn
streamlit
pdfplumber
python-docx
spacy
sentence-transformers
torch
transformers
spacy
spacy[en]
